Fixing the code

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -2,5 +2,6 @@
 
 exec { 'killmenow':
   command  => 'pkill -i',
-  provider => 'shell',
+  provider => shell,
+  onlyif   => 'killmenow',
 }
